- UI almost ready, missing the error message, and total amount on top of bottom bar
