+++ app/src/main/java/com/kickstarter/ui/activities/compose/projectpage/ProjectPledgeButtonAndFragmentContainer.kt
@@ -257,7 +257,8 @@ fun ProjectPledgeButtonAndFragmentContainer(
                                         modifier = Modifier,
                                         environment = environment ?: Environment.builder().build(),
                                         lazyColumnListState = rememberLazyListState(),
-                                        rewardItems = addOns,
+                                        selectedReward = selectedReward ?: Reward.builder().build(),
+                                        addOns = addOns,
                                         project = project,
                                         onItemAddedOrRemoved = onAddOnAddedOrRemoved,
                                         onContinueClicked = onAddOnsContinueClicked,
